+++ src/app/collection/collection.component.ts
@@ -65,7 +65,7 @@ export class CollectionComponent implements OnInit {
     addDashboard() {
         const dialogRef = this.dialog.open(DashboardDialogComponent, {
             width: '600px',
-            data: {dashboard: {collection_id: this.collection.id, users: []}}
+            data: {dashboard: {collection_id: this.collection.id, users: [], variables: []}}
         });
 
         dialogRef.afterClosed().subscribe(result => {